WildcardMatcher: treat unmatched filename prefixes as no match
This change fixes the following
RepositoryTest::testDefinitionByExtension and
RepositoryTest::testDefinitionsForFileName data rows:
not-Makefile.dic
*qmakefile.cpp
*_makefile.mm

The performance of Repository::definitionForFileName() and
Repository::definitionsForFileName() is practically unaffected by this
fix.

Rename test files: highlight.debian.control* to control*. This is
necessary for tests to pass with the fix. The Debian Control pattern is
"control" - a literal pattern. It does not match the filename
"highlight.debian.control". The tests passed before because of the bug
fixed in this commit.