Replace "SConscript" with ".profile" to fix a test failure
The Python-Kig language must be installed along with Kig:
https://commits.kde.org/kig?path=scripting/python-kig.xml

This language is absent from syntax-highlighting repository, and so the
test fails on the CI system where Kig is not installed.

The replacement literal pattern - ".profile" - has the same properties
as "SConscript", but is present in two files - data/syntax/bash.xml and
data/syntax/zsh.xml - in syntax-highlighting repository.