Port AbstractHighlighter::highlightLine to QStringView
If your line is already a QString this makes no difference, otherwise this
now allows to perform highlighting without the need of splitting lines into
QString copies.

The old QString variant has now deprecation warning, as the removal of that
will automatically "port" user code. With that overload disabled there is
one difference though that can be seen in the test changes: QLatin1String
cannot be converted implicitly to a QStringView. For production code this
shouldn't matter though.