install a index.katesyntax when QRC_SYNTAX is OFF
The index file allows to gather all the metadata of the installed syntaxes in a single file.

Without it, all installed xml files will be partially read. This is a bit slow on HHD or other devices such as SD cards.

The following reads generally benefit from the disk cache and are less prone to slowness.