New Folding for markdown fenced code
While the new style does include a significant amount of duplication,
the highlighting in editors like kate feels much more natural. The old
problem was that, due to the `beginRegion` being in a lookahead regex,
the size of the `beginfold` was always 0 characters long, but `endfold`
was 3 or more characters long, leading to a very uneven and irritating
highlighting.