R: add numeric suffix (i and L)
BUG: 468394

It doesn't seem to be a bug on our side, but I did make one change: highlighting suffixes.

I wonder if we shouldn't generalize and standardize this? There are several behaviors depending on the syntax:

There are several behaviors depending on the syntax:

- no treatment for suffixes, they often (always?) have the Normal color
- number and suffix have the same color, but the color of the same suffix can vary between Float, Int, Hex, etc.
- suffixes have a dedicated color

As shown in this commit, I'm in favor of the dedicated color and propose to generalize it to all syntaxes, what do you think?