Alternate Names for Definitions
This use case comes from NeoChat where we've started using `ksyntaxhighligher` to highlight code in messages. The difficulty we're having is that typically when using a markdown code block the convention is to use lower case and in some cases abbreviations or slightly different names, e.g. cpp instead of C++ or hs instead of Haskell.

This means in order to handle things efficiently we need to be able to handle two cases:
- Case Insensitive search on the name as we would expect `rust` rather than `Rust`
- Alternate names where these are different to the main definition name

This adds MR adds both as it seems best to do this in the library rather than go it alone in NeoChat