Add Syntax Highlighting for Vue Template Files
Vue is a common JS framework.

https://vuejs.org/

Vue template files are similar to HTML, but with a few differences.

This highlighter aims to better accommodate those differences.

It also gives us an explicit language name for use with Kate's LSP.

So we can add a sensible default LSP config for Kate in future.

Signed-off-by: James Zuccon <zuccon@gmail.com>

@teams/qa