Refactor `alerts` keywords to `crash keywords` and add `import keyword` list
I tink `assert`, `panic` and `todo` should be highlighted differently because
they can result in a crash at runtime. They were styled as `dsAlert` before wich
is used for comments not keywords, I did not see any default style better suited
than `dsControlFlow`. It does not fit as well semantically but it is unlikely to
be used elsewhere since Gleam has a very small number of control flow keywords
(`case` and `if`).