replace DefinitionRef type with DefinitionData* in immediateIncludedDefinitions
This change slightly speeds up syntax loading and the call to includedDefinitions().

immediateIncludedDefinitions contains only valid definitions data pointer.
The only way to invalidate them is to delete the repository or perform a reload().
At this point, all definitions become invalid and checking isValid()
at the start of Definition::includedDefinitions() is sufficient.

Moreover, as the inclusion of other syntax is an infrequent occurrence,
immediateIncludedDefinitions is now a QVarLengthArray of 4 values on a
stack avoids allocations 98% of the time.

Following these changes, DefinitionRef is no longer used.