new feature: stacking several contexts at the same time
The new syntax `context="name1!name2##SH!name3"` allows multiple contexts
to be stacked.

This syntax avoids the multiplication of contexts when a common syntax
precedes a particular syntax.

For example, strings (single quote or double quote)
in () or [] require the definition of four string contexts:

- single quote + ]
- double quote + ]
- single quote + )
- double quote + )

Which is represented by the following graph:

```
"[" -> SqStr_1 -> "]"
    \> DqStr_1 /
"(" -> SqStr_2 -> ")"
    \> DqStr_2 /
```

By stacking the closure before SqStr or DqStr, it is possible
to reduce the number of contexts by jumping over the closure via a `#pop`.

```
"[" -> SqStr -> "]"
    \> DqStr /
"(" -> SqStr -> ")"
    \> DqStr /
```

If a capture is present (e.g. `<RegExpr String="(...)"/>`),
it propagates to all contexts.

`utils/generate-dit-file.py` is updated accordingly.